anime_dance_social_media/cloud_run/full_pipeline/requirements.txt
# Cloud Run Full Pipeline Service - Dependencies

# Web Framework
flask==3.0.0
gunicorn==21.2.0

# Google Cloud
google-cloud-storage==2.10.0
google-cloud-firestore==2.13.0
google-auth==2.23.0

# Video Processing
moviepy==2.0.0
opencv-python==4.8.1.78
Pillow==10.1.0

# AI/ML APIs
google-generativeai==0.3.0
requests==2.31.0

# Utilities
python-dotenv==1.0.0
uuid==1.30

# Production
Werkzeug==3.0.1
